Feedback changes on PR resolved

+++ admin/class-purger.php
@@ -108,15 +108,11 @@ abstract class Purger {
 			case 'spam':
 			case 'unapproved':
 			case 'trash':
-				if ( 'approve' === $oldstatus ) {
+				if ( 'approve' === $oldstatus && 1 === (int)$nginx_helper_admin->options['purge_page_on_deleted_comment'] ) {
 
-					if ( 1 === (int)$nginx_helper_admin->options['purge_page_on_deleted_comment'] ) {
-
-						$this->log( '* Comment ( ' . $_comment_id . ' ) removed as ( ' . $newstatus . ' ). Post ( ' . $_post_id . ' ) purging...' );
-						$this->log( '* * * * *' );
-						$this->purge_post( $_post_id );
-
-					}
+					$this->log( '* Comment ( ' . $_comment_id . ' ) removed as ( ' . $newstatus . ' ). Post ( ' . $_post_id . ' ) purging...' );
+					$this->log( '* * * * *' );
+					$this->purge_post( $_post_id );
 
 				}
 				break;
